pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "hackingBuddyGPT"
# original author was Andreas Happe, for an up-to-date list see
# https://github.com/ipa-lab/hackingBuddyGPT/graphs/contributors
authors = [
	{ name = "HackingBuddyGPT maintainers", email = "maintainers@hackingbuddy.ai" }
]
maintainers = [
	{ name = "Andreas Happe", email = "andreas@offensive.one" },
	{ name = "Juergen Cito", email = "juergen.cito@tuwien.ac.at" }
]
description = "Helping Ethical Hackers use LLMs in 50 lines of code"
readme = "README.md"
keywords = ["hacking", "pen-testing", "LLM", "AI", "agent"]
requires-python = ">=3.10"
version = "0.4.0"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
	"Development Status :: 4 - Beta",
]
dependencies = [
    'fabric == 3.2.2',
    'Mako == 1.3.2',
    'requests == 2.32.3',
    'rich == 13.7.1',
    'tiktoken == 0.8.0',
    'instructor == 1.7.2',
    'PyYAML == 6.0.1',
    'python-dotenv == 1.0.1',
    'pypsexec == 0.3.0',
    'pydantic == 2.8.2',
    'openai == 1.65.2',
    'google-generativeai >= 0.3.0',
    'BeautifulSoup4',
    'nltk',
    'fastapi == 0.114.0',
    'fastapi-utils == 0.7.0',
    'jinja2 == 3.1.6',
    'uvicorn[standard] == 0.30.6',
    'dataclasses_json == 0.6.7',
    'websockets == 13.1',
    'langchain-community',
    'langchain-openai',
    'markdown',
    'chromadb',
    'langchain-chroma',
]

[project.urls]
Homepage = "https://www.hackingbuddy.ai"
Documentation = "https://docs.hackingbuddy.ai"
Repository = "https://github.com/ipa-lab/hackingBuddyGPT"
"Bug Tracker" = "https://github.com/ipa-lab/hackingBuddyGPT/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"hackingBuddyGPT.usecases.privesc.templates" = ["*.txt"]
"hackingBuddyGPT.usecases.examples" = ["*.txt"]

[tool.pytest.ini_options]
pythonpath = "src"
addopts = ["--import-mode=importlib"]
[project.optional-dependencies]
testing = ['pytest', 'pytest-mock']
dev = [
	'ruff',
]
rag-usecase = [
    'langchain-community',
	'langchain-openai',
	'markdown',
	'chromadb',
	'langchain-chroma',
	'google-generativeai >= 0.3.0',
]

[project.scripts]
wintermute = "hackingBuddyGPT.cli.wintermute:main"
hackingBuddyGPT = "hackingBuddyGPT.cli.wintermute:main"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = ["E501", "F401", "F403"]
